requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-writable-nested==0.6.3
pytz==2021.1
sqlparse==0.4.2
